requirements/local.txt
asgiref==3.4.1
Django==4.0.1
django-ckeditor==6.2.0
django-js-asset==1.2.2
Pillow==9.0.0
psycopg2==2.9.3
sqlparse==0.4.2
tzdata==2021.5